(#17) 🚀 feat: 회원가입 이벤트 구현
